改浮点

+++ BubbleKnight/Assets/BubbleKnight/Scripts/UIManager.cs
@@ -47,13 +47,13 @@ public class UIManager : MonoBehaviour
         }
     }
 
-    public void SetSpeed(int speed)
+    public void SetSpeed(float speed)
     {
         speedUI.GetComponent<TextMeshProUGUI>().text = "Speed: " + speed + " m/s";
     }
 
     public void SetTarget(float now, float target)
     {
-        targetUI.GetComponent<TextMeshProUGUI>().text = now + " m" + " / " + target + " m";
+        targetUI.GetComponent<TextMeshProUGUI>().text = (int)now + " m" + " / " + (int)target + " m";
     }
 }
